requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.4.0+cu121
torchvision==0.19.0+cu121
numpy
matplotlib
Pillow
opencv-python
tqdm
scikit-learn
pretrainedmodels
pandas